+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/jdbc/template/OptionalQueryProvider.java
@@ -1,8 +1,0 @@
-package com.dwarfeng.subgrade.sdk.jdbc.template;
-
-/**
- * @author DwArFeng
- * @since 1.1.0
- */
-public interface OptionalQueryProvider {
-}